requirements.txt
click
ib_insync
loguru
pandas >=1.3.0,<1.4
numpy
pydantic
python-dotenv
pytz
requests
tqdm
